+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ### Added
 - Add PHP 8.2 support
+- Add information to the Laravel about command.
 
 ### Changed
 - Switch from `getHash` to `getHashAndIgnoreFailures` in the HashLoader to iterate all finders on a failure.